feat: Introduce tool to verify a storage module
when the `verify` launch option is set:
1. the node will launch from local state without joining the network
2. it will then iterate through all storage_modules validating different
   indices and proofs
3. any chunk that fails validation will be flagged as invalid
4. the next time the node is launchd with syncing enabled it will try to
   sync and pack the flagged chunks

This change also refactors the code for validating proofs. It introduces
a #chunk_proof record which could eventually replace the tuples and
maps that are currently used when dealing with proofs. This is just
an initial step, though, and the chunk_proof record is only used
internally (i.e. no updates to any serialized data structures)