feat: add support for an e2e test mode
Tests run in e2e mode will more closely match the normal node
operation. Notably: full randomx and normal poa1 vs. poa2 difficulty.

There are still a lot of constants that are reset in this mode -
*mostly* to speed up the test run and allow it to work on smaller
blockchains. But if material differences are discovered we can
reset some of those constants back to more representative values
as needed

This commit also changes the DEBUG constant to TEST. We have
been using ifdef(DEBUG) to give different values to constants
in test-mode vs. in normal mode. However some of our 3rd party
dependencies use DEBUG to signify builds which include debugging
information. We also use the `debug` command line option to add
debugging information. The overloading of DEBUG has been a point
of confusion, so just changing it to TEST to clarify the difference