Improve ar_tx_emitter performance
Only current (active) peers should be able to receive the
chunks, others will generate unnecessary timeouts. In fact
using the active peers from the last 24 hours can also be
a great solution if we assume a peer seen during the last
24 hours will probably be come back quick.

When chunks are propagated across the network, the to be
acknowledged are the trusted peer. On the previous version,
the trusted peer where put at the end of the list.

Cleaned up ar_tx_emitter_sup, converting old tuple to map.

Refacto and added comments and documentation.

see: https://github.com/ArweaveTeam/arweave-dev/issues/720