Fix Github Workflow Test Issue (#731)
Following the modification of the entry-point, the default shell
changed. Instead of using bash we are now using sh (dash). The special
flag failpipe is not supported and then, the latest command to be
executed in a pipe is valid if it does not fail.

Created a canary test in the pipeline that always fail to avoid
this kind of issue in the future.

github_workflow.sh script has been rollbacked. The new entry-point
was not correctly working with it.

see: https://github.com/ArweaveTeam/arweave-dev/issues/816