Fix Slow Node Shutdown (#739)
This commit should fix the issue when a node is shutting down
and the connections are still up/not correctly closed. It seems
the default behavior is to wait for all connections to be
properly closed, but this is not what we want, even more
if clients are requesting huge data with slow connections.

see: https://github.com/ArweaveTeam/arweave-dev/issues/817