fix: fix bottleneck in ar_global_sync_record
When a lot of chunks are being added to the global sync record
(e.g. during a sync, pack, or repack), it needs to check for
replica_2_9 chunks a lot. Previously each check used the
ar_storage_module:get_by_id(StoreID) call which inclues a couple
ets calls. At high volume this was a bottleneck which eventually
caused the ar_global_sync_record message queue to back up and OOM.

Fix is to have the sync_record events include the StorageModule
rather than the StoreID and so bypass the get_by_id call