Enable VDF tests on MacOS (#805)
This commit enables VDF tests for MacOS. because MacOS
does not support containers at this time, some changes
were required. Indeed, all runners are isolated using
a dedicated user and environment, but rebar3 is using
symlinks when dealing with dependencies. To avoid issues,
links are deferenced. Artifacts will be bigger, but it
is possible to run more than one worker at a time now,
in a distributed environment.

The cache has been disabled (for now). When enabled,
the wrong version of cmake is taken, and it blocks
the compilation.

Fixed duplicated code in standard tests.

see: https://github.com/ArweaveTeam/arweave-dev/issues/880

update