Add support for Erlang 26 and MacOS (#835)
Arweave is working correctly on R26 and the next release
should offer a version of it running on this Erlang version.

Create makefile to help releasing on MacOS, only compatibe
with MacOS for the moment. Because MacOS does not have an
easy way to create a chroot, a container or a jail, the
only solution right now is to create a system from
scratch everytime we need to create a release. This is a
first step.

see: https://github.com/ArweaveTeam/infra/issues/191