fix: ar_entropy_cache is no longer a gen_server
Since it is primarily a wrapper around access to an ETS table the gen_server layer was causing some race conditions. Now that it's a striaght interface to ETS we can rely on ETS to resolve certain classes of race conditins (e.g. what happened before is a `put` was implemented as a `cast` and before the `cast` was processed a straight read from the ETS table was done and missed the put update)